TERF-83: Change int32 to int64 for uid_or_gid

+++ codegen/latest/model_share_acl.go
@@ -9,10 +9,10 @@
 package swagger
 
 type ShareAcl struct {
-	Grantee string `json:"grantee,omitempty"`
+	Grantee     string `json:"grantee,omitempty"`
 	Permissions string `json:"permissions,omitempty"`
-	SidStr string `json:"sid_str,omitempty"`
-	UidOrGid int32 `json:"uid_or_gid,omitempty"`
-	Name string `json:"name,omitempty"`
-	Fqdn string `json:"fqdn,omitempty"`
+	SidStr      string `json:"sid_str,omitempty"`
+	UidOrGid    int64  `json:"uid_or_gid,omitempty"`
+	Name        string `json:"name,omitempty"`
+	Fqdn        string `json:"fqdn,omitempty"`
 }
